fix: injection split start
